+++ critical-logger.php
@@ -3,7 +3,7 @@
  * Plugin Name: Critical Event Logger
  * Plugin URI: https://github.com/fly380/critical-event-logger
  * Description: Логування критичних подій із швидким AJAX-переглядом, парсером «склеєних» рядків, частотністю IP, Geo/пул-визначенням, ручним блокуванням (.htaccess для Apache 2.2/2.4), ротацією й очищенням логів, GeoBlock та опційними AI-інсайтами.
- * Version: 1.1
+ * Version: 1.1.1
  * Author: Казмірчук Андрій
  * Author URI: https://www.facebook.com/fly380/
  * Text Domain: fly380
@@ -289,6 +289,7 @@ if ( ! function_exists('crit_fix_changelog_spacing_js') ) {
 /* Підключаємо основні файли плагіна */
 require_once plugin_dir_path(__FILE__) . 'logger.php';
 require_once plugin_dir_path(__FILE__) . 'logger-hooks.php';
+require_once plugin_dir_path(__FILE__) . 'bots.php';
 if (file_exists(plugin_dir_path(__FILE__) . 'privacy.php')) {
 	require_once plugin_dir_path(__FILE__) . 'privacy.php';
 }